Implement 2nd motor and fix timing
Implement the 2nd motor in the pair by switching the target position and
current position to be tracked by the pair instead of each motor
individually. Resolve timing issue by fixing wait logic for motor speed,
as it was being calculated improperly before. Also remove delay that was
always occurring and messing with the loop, and replace it with a delay
that only happens when the motor does not move (10ms).