Mostly working changes to accel/decel
Mostly working changes to acceleration and deceleration algorithms. It
does a weird thing were it decelerates a bit and then accelerates, so
there's some unintended oscillation. But for the most part it is
working. The acceleration is working fine (and actually as it was
intended, since it wasn't before). But the decelartion isn't